+++ core/report/service_level.go
@@ -6,12 +6,12 @@ import (
 	"github.com/reliablyhq/cli/core/manifest"
 )
 
-func getCurrentAvailability(m *manifest.Manifest, d time.Duration) (float32, error) {
-	return 100 * r.Float32(), nil
+func getCurrentAvailability(m *manifest.Manifest, d time.Duration) (float64, error) {
+	return 100 * r.Float64(), nil
 }
 
-func getCurrentErrorPc(m *manifest.Manifest, d time.Duration) (float32, error) {
-	return 25 * r.Float32(), nil
+func getCurrentErrorPc(m *manifest.Manifest, d time.Duration) (float64, error) {
+	return 25 * r.Float64(), nil
 }
 
 func get99PercentLatency(m *manifest.Manifest, d time.Duration) (time.Duration, error) {
